recipe-miner/requirements.txt
# Recipe Miner Dependencies
google-generativeai>=0.3.2
pydantic>=2.5.0
duckduckgo-search>=4.1.0
trafilatura>=1.6.4
supabase>=2.3.0
python-dotenv>=1.0.0
tenacity>=8.2.3
